separate out the search bar into its own component. add OracleEmbedContext
