+++ lib/components/oracle/embed/OracleEmbedContext.ts
@@ -1,0 +1,14 @@
+import { createContext } from "react";
+import { OracleSearchBarManager } from "./search-bar/oracleSearchBarManager";
+
+interface OracleEmbedContext {
+  apiEndpoint: string;
+  token: string;
+  searchBarManager: OracleSearchBarManager;
+}
+
+export const OracleEmbedContext = createContext<OracleEmbedContext>({
+  apiEndpoint: "",
+  token: "",
+  searchBarManager: OracleSearchBarManager(),
+});
